tests/wpt/mozilla/tests/accessibility-tree/accessibility-update-text-change.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<section>
  <h1 id="h1">This is an h1</h1>
</section>

<script>
  setup({ explicit_done: true });
  window.onload = () => {
    test((t) => {
      ServoTestUtils.ensureAccessibilityActive();
      let initial = ServoTestUtils.forceAccessibilityUpdate();

      assert_equals(initial.nodesUpdatedFromDom, 16, "nodesUpdatedFromDom");
      assert_equals(initial.nodesUpdatedFromTree, 16, "nodesUpdatedFromTree");
      assert_equals(initial.nodesInTreeUpdate, 16, "nodesInTreeUpdate");

      h1.firstChild.appendData(", now with more text");
      let update = window.ServoTestUtils.forceAccessibilityUpdate();

      // Just the text node has changed
      assert_equals(update.nodesUpdatedFromDom, 1, "nodesUpdatedFromDom");

      // The document, <body>, <section>, <h1> and the text node.
      assert_equals(update.nodesUpdatedFromTree, 5, "nodesUpdatedFromTree");

      //  updated nodes:
      // - <h1>, computed label changed
      // - text child of <h1>, value changed
      assert_equals(update.nodesInTreeUpdate, 2, "nodesInTreeUpdate");
    }, "Append data to text node should update 2 nodes");

    done();
  };
</script>
